First (signed) commit
of installer for the develop branch of Zim, which contains the new
"plugin mechanism". See zimfw/zimfw#323

The code of this installer was based on the code of the new zimfw
utility, and more specifically on the install.zsh and update.zsh
scripts, developed for the new Zim "plugin mechanism".
Keeping the same nice (while still minimalistic) output style too.