Updated install for single-file Zim script
So the zimfw repo is not cloned anymore, but instead, a zimfw.zsh script
from that repo is directly downloaded with either wget or curl.

Also the templates that were in the zimfw repo are now moved to this
one, embedded into the install script. The install script is also a
single-file one, following the same approach used in the zimfw.zsh
script.

In a effort to help making the install script maintainable, I've broken
it into small ERB templates. This also adds the ability to pre-process
the Zsh code with Ruby code. To build the script, use `make`.